refactor(api): split errors.rs into error/ (problem + classify); ApiError #[non_exhaustive]
Decomposes the 759-line god-file crates/api/src/errors.rs into a focused
error/ module per canon §12.7 and spec §6.1:

- error/problem.rs — ProblemDetails + ValidationFieldError (RFC 9457 wire
  type; field names, serde annotations, and application/problem+json
  content-type header unchanged — public contract preserved byte-identical).
- error/classify.rs — From<ValidationError>, From<ValidationErrors>,
  From<PermissionDenied> conversions + workflow_error_pointer helper.
- error/mod.rs — ApiError enum (now #[non_exhaustive] per STYLE §5 / spec
  D6), ApiResult alias, IntoResponse impl, and the five unit tests moved
  from the old file.

lib.rs: pub mod errors → pub mod error; re-exports updated to include
ProblemDetails. No shim added (grep confirmed zero external nebula_api::errors::
references in apps/, examples/, or tests/).

Updated 21 internal use sites (handlers/, middleware/, services/, auth/,
extractors/, state.rs, tests/common/mod.rs) from crate::errors:: to
crate::error::. No test dropped; nextest: 236 passed, 1 skipped.
